mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.5",
]
